finserv_lakehouse/app/requirements.txt
databricks-sdk>=0.20.0
pandas>=2.0.0
